Removed utilities file
Reverse was not used and the other two functions were
only used once so it makes sense to keep them in the
file that they are needed.

New files in the same package should be avoided if at all
possible. They make code a lot harder to understand.
